add simple host discovery
On begin a UDP client is created that periodically broadcasts the marco message and it sets up a UDP server that listens on same port for a response from a host. The response as of yet it not acted upon.

A change to Cinder-Asio's UdpServer is required to enable address reuse, a solution and discussion follows in issue #5 of the main repo.