handle multiple messages from TCP server read event handler
The message-listening socket can receive multiple messages at once under some circumstances, specifically the one I'm seeing is on OS X Mavericks, AppNap will coalesce the application's timers and in turn, the Boost.Asio / network events coming into the app are queued a bit - this could be a Cinder thing, as maybe it doesn't do the right thing with event pump or it could just be the default OS behavior, I'm not really sure. Regardless, multiple messages can be received at once so the data is more carefully parsed and the messages pulled out and processed from it.